day1: part1 and part2 (rec and non rec versions)

+++ day7/include/day7.h
@@ -18,8 +18,8 @@
 
 typedef struct s_eq
 {
-	int	test_val;
-	Array operands;
+	long long	test_val;
+	Array		operands;
 }	t_eq;
 
 typedef struct s_info
